Update project name.
        So Richard and I decided to try to catch [the small shark].
With a great deal of strategy and effort and shouting, we managed to
maneuver the shark, over the course of about a half-hour, to a sort of
corner of the lagoon, so that it had no way to escape other than to
flop up onto the land and evolve.  Richard and I were inching toward
it, sort of crouched over, when all of a sudden it turned around and --
I can still remember the sensation I felt at that moment, primarily in
the armpit area -- headed right straight toward us.
        Many people would have panicked at this point.  But Richard and
I were not "many people."  We were experienced waders, and we kept our
heads.  We did exactly what the textbook says you should do when you're
unarmed and a shark that is nearly two feet long turns on you in water
up to your lower calves: We sprinted I would say 600 yards in the
opposite direction, using a sprinting style such that the bottoms of
our feet never once went below the surface of the water.  We ran all
the way to the far shore, and if we had been in a Warner Brothers
cartoon we would have run right INTO the beach, and you would have seen
these two mounds of sand racing across the island until they bonked
into trees and coconuts fell onto their heads.
                -- Dave Barry, "The Wonders of Sharks on TV"