Remove unused sprite
"In corporate life, I think there are three important areas which contracts
can't deal with, the area of conflict, the area of change and area of reaching
potential.  To me a covenant is a relationship that is based on such things
as shared ideals and shared value systems and shared ideas and shared
agreement as to the processes we are going to use for working together.  In
many cases they develop into real love relationships."
-- Max DePree, chairman and CEO of Herman Miller Inc., "Herman Miller's
   Secrets of Corporate Creativity", The Wall Street Journal, May 3, 1988