Convert to TypeScript.
To compile : just run tsc.

Inglish Spocken Hier: some mangled translations

        Sign on a cabin door of a Soviet Black Sea cruise liner:
                Helpsavering apparata in emergings behold many whistles!
                Associate the stringing apparata about the bosums and meet
                behind, flee then to the indifferent lifesaveringshippen
                obedicing the instructs of the vessel.

        On the door in a Belgrade hotel:
                Let us know about any unficiency as well as leaking on
                the service. Our utmost will improve it.

                -- Colin Bowles

Even the clearest and most perfect circumstantial evidence is likely to be at
fault, after all, and therefore ought to be received with great caution.  Take
the case of any pencil, sharpened by any woman; if you have witnesses, you will
find she did it with a knife; but if you take simply the aspect of the pencil,
you will say that she did it with her teeth.
                -- Mark Twain, "Pudd'nhead Wilson's Calendar"

Catch a wave and you're sitting on top of the world.
                -- The Beach Boys